[Bronze III] Title: Simple Arithmetic, Time: 0 ms, Memory: 2020 KB, Score: 75 point -BaekjoonHub

+++ 백준/Bronze/15858. Simple Arithmetic/Simple Arithmetic.cc
@@ -7,7 +7,7 @@ int main()
     ios::sync_with_stdio(false);
     cin.tie(nullptr);
 
-    double a, b, c; cin >> a >> b >> c;
+    long double a, b, c; cin >> a >> b >> c;
     cout << fixed << setprecision(16);
     cout << a * b / c;
 }
